Added new constructor
